refactor(clients): shrink GitHub client wrappers to narrow boundaries (#406)
Split monolithic GitHubClient into narrow use-case ports:
- Created 7 narrow port protocols (Auth, PRRead, PRWrite, PRDiff, PRComment, IssueRead, IssueWrite)
- Reorganized mixins by responsibility (read/write)
- Reduced github_pr_ops.py from 520 LOC to 21 LOC (re-export module)
- Migrated pr_analysis_service to direct gh CLI calls
- Added architectural guidance in docs/standards/client-boundaries.md

Verification:
- All 441 tests pass
- Type check clean (mypy)
- Lint clean (ruff)
- All files under 400 LOC

Closes #406
