+++ src/vibe3/services/pr_analysis_service.py
@@ -3,6 +3,7 @@
 Provides reusable PR analysis logic for both commands and services.
 """
 
+import subprocess
 from pathlib import Path
 from typing import cast
 
@@ -184,14 +185,29 @@ def _calculate_risk_score(
 
 
 def _get_recent_commits(pr_number: int, limit: int = 5) -> list[CommitInfo]:
-    """Return latest commit messages for a PR."""
-    from vibe3.clients.github_client import GitHubClient
+    """Return latest commit messages for a PR via direct gh CLI call."""
     from vibe3.utils.git_helpers import get_commit_message
 
-    gh = GitHubClient()
-
     try:
-        commit_shas = gh.get_pr_commits(pr_number)
+        # Direct gh CLI call instead of client wrapper
+        result = subprocess.run(
+            [
+                "gh",
+                "pr",
+                "view",
+                str(pr_number),
+                "--json",
+                "commits",
+                "--jq",
+                ".commits[].oid",
+            ],
+            capture_output=True,
+            text=True,
+            check=True,
+        )
+        commit_shas = [
+            line.strip() for line in result.stdout.strip().split("\n") if line.strip()
+        ]
     except Exception as e:
         logger.warning(f"Failed to get commits for PR {pr_number}: {e}")
         return []
@@ -217,13 +233,27 @@ def _get_recent_commits(pr_number: int, limit: int = 5) -> list[CommitInfo]:
 
 
 def _get_pr_commit_count(pr_number: int) -> int:
-    """Return total commit count for a PR."""
-    from vibe3.clients.github_client import GitHubClient
-
-    gh = GitHubClient()
-
+    """Return total commit count for a PR via direct gh CLI call."""
     try:
-        commits = gh.get_pr_commits(pr_number)
+        # Direct gh CLI call instead of client wrapper
+        result = subprocess.run(
+            [
+                "gh",
+                "pr",
+                "view",
+                str(pr_number),
+                "--json",
+                "commits",
+                "--jq",
+                ".commits[].oid",
+            ],
+            capture_output=True,
+            text=True,
+            check=True,
+        )
+        commits = [
+            line.strip() for line in result.stdout.strip().split("\n") if line.strip()
+        ]
         return len(commits)
     except Exception as e:
         logger.warning(f"Failed to get commit count for PR {pr_number}: {e}")
