+++ tests/vibe3/commands/test_inspect_commits.py
@@ -17,17 +17,14 @@ from vibe3.exceptions import GitError
 def test_get_recent_commits_success():
     """Successfully get recent commits."""
     with (
-        patch("vibe3.clients.github_client.GitHubClient") as mock_github_client,
+        patch("vibe3.services.pr_analysis_service.subprocess.run") as mock_run,
         patch("vibe3.utils.git_helpers.get_commit_message") as mock_get_commit_message,
     ):
-        # Mock get_pr_commits
-        mock_gh = MagicMock()
-        mock_gh.get_pr_commits.return_value = [
-            "abc123",
-            "def456",
-            "ghi789",
-        ]
-        mock_github_client.return_value = mock_gh
+        # Mock subprocess.run to return commit SHAs
+        mock_run.return_value = MagicMock(
+            returncode=0,
+            stdout="abc123\ndef456\nghi789\n",
+        )
 
         # Mock get_commit_message
         mock_get_commit_message.side_effect = [
@@ -48,12 +45,13 @@ def test_get_recent_commits_success():
 def test_get_recent_commits_limit():
     """Respect limit parameter."""
     with (
-        patch("vibe3.clients.github_client.GitHubClient") as mock_github_client,
+        patch("vibe3.services.pr_analysis_service.subprocess.run") as mock_run,
         patch("vibe3.utils.git_helpers.get_commit_message") as mock_get_commit_message,
     ):
-        mock_gh = MagicMock()
-        mock_gh.get_pr_commits.return_value = ["a", "b", "c", "d", "e", "f"]
-        mock_github_client.return_value = mock_gh
+        mock_run.return_value = MagicMock(
+            returncode=0,
+            stdout="a\nb\nc\nd\ne\nf\n",
+        )
         mock_get_commit_message.return_value = "message"
         result = _get_recent_commits(42, limit=3)
         assert len(result) == 3
@@ -61,20 +59,19 @@ def test_get_recent_commits_limit():
 
 def test_get_recent_commits_empty():
     """Handle empty commits list."""
-    with patch("vibe3.clients.github_client.GitHubClient") as mock_github_client:
-        mock_gh = MagicMock()
-        mock_gh.get_pr_commits.return_value = []
-        mock_github_client.return_value = mock_gh
+    with patch("vibe3.services.pr_analysis_service.subprocess.run") as mock_run:
+        mock_run.return_value = MagicMock(
+            returncode=0,
+            stdout="\n",
+        )
         result = _get_recent_commits(42)
         assert result == []
 
 
 def test_get_recent_commits_github_error():
     """Handle GitHub API error gracefully."""
-    with patch("vibe3.clients.github_client.GitHubClient") as mock_github_client:
-        mock_gh = MagicMock()
-        mock_gh.get_pr_commits.side_effect = Exception("API error")
-        mock_github_client.return_value = mock_gh
+    with patch("vibe3.services.pr_analysis_service.subprocess.run") as mock_run:
+        mock_run.side_effect = Exception("API error")
         result = _get_recent_commits(42)
         assert result == []
 
@@ -82,12 +79,13 @@ def test_get_recent_commits_github_error():
 def test_get_recent_commits_git_error():
     """Skip commits with git errors, continue with others."""
     with (
-        patch("vibe3.clients.github_client.GitHubClient") as mock_github_client,
+        patch("vibe3.services.pr_analysis_service.subprocess.run") as mock_run,
         patch("vibe3.utils.git_helpers.get_commit_message") as mock_get_commit_message,
     ):
-        mock_gh = MagicMock()
-        mock_gh.get_pr_commits.return_value = ["abc123", "def456", "ghi789"]
-        mock_github_client.return_value = mock_gh
+        mock_run.return_value = MagicMock(
+            returncode=0,
+            stdout="abc123\ndef456\nghi789\n",
+        )
         # Second commit fails
         mock_get_commit_message.side_effect = [
             "Add feature X",
@@ -104,12 +102,13 @@ def test_get_recent_commits_git_error():
 def test_get_recent_commits_short_sha():
     """SHA is shortened to 7 characters."""
     with (
-        patch("vibe3.clients.github_client.GitHubClient") as mock_github_client,
+        patch("vibe3.services.pr_analysis_service.subprocess.run") as mock_run,
         patch("vibe3.utils.git_helpers.get_commit_message") as mock_get_commit_message,
     ):
-        mock_gh = MagicMock()
-        mock_gh.get_pr_commits.return_value = ["abcdefghijklmnopqrstuvwxyz123456"]
-        mock_github_client.return_value = mock_gh
+        mock_run.return_value = MagicMock(
+            returncode=0,
+            stdout="abcdefghijklmnopqrstuvwxyz123456\n",
+        )
         mock_get_commit_message.return_value = "message"
         result = _get_recent_commits(42)
         assert result[0]["sha"] == "abcdefg"  # First 7 chars
@@ -120,29 +119,29 @@ def test_get_recent_commits_short_sha():
 
 def test_get_pr_commit_count_success():
     """Get commit count successfully."""
-    with patch("vibe3.clients.github_client.GitHubClient") as mock_github_client:
-        mock_gh = MagicMock()
-        mock_gh.get_pr_commits.return_value = ["a", "b", "c", "d", "e"]
-        mock_github_client.return_value = mock_gh
+    with patch("vibe3.services.pr_analysis_service.subprocess.run") as mock_run:
+        mock_run.return_value = MagicMock(
+            returncode=0,
+            stdout="a\nb\nc\nd\ne\n",
+        )
         result = _get_pr_commit_count(42)
         assert result == 5
 
 
 def test_get_pr_commit_count_empty():
     """Handle empty commits."""
-    with patch("vibe3.clients.github_client.GitHubClient") as mock_github_client:
-        mock_gh = MagicMock()
-        mock_gh.get_pr_commits.return_value = []
-        mock_github_client.return_value = mock_gh
+    with patch("vibe3.services.pr_analysis_service.subprocess.run") as mock_run:
+        mock_run.return_value = MagicMock(
+            returncode=0,
+            stdout="\n",
+        )
         result = _get_pr_commit_count(42)
         assert result == 0
 
 
 def test_get_pr_commit_count_error():
     """Return 0 on error."""
-    with patch("vibe3.clients.github_client.GitHubClient") as mock_github_client:
-        mock_gh = MagicMock()
-        mock_gh.get_pr_commits.side_effect = Exception("API error")
-        mock_github_client.return_value = mock_gh
+    with patch("vibe3.services.pr_analysis_service.subprocess.run") as mock_run:
+        mock_run.side_effect = Exception("API error")
         result = _get_pr_commit_count(42)
         assert result == 0
